+++ frontend/src/views/Login/index.tsx
@@ -1,3 +1,3 @@
 import LogIn from './LogIn';
 
-export default LogIn; 
+export default LogIn;
