+++ frontend/src/views/Registration/CardForm/index.tsx
@@ -1,0 +1,3 @@
+import CardForm, { Props } from './CardForm';
+
+export default CardForm;
