+++ frontend/src/views/Registration/CardForm/styles.module.scss
